requirements.txt
neo4j-graphrag[openai]
neo4j
openai
requests
streamlit
python-dotenv
